Backup VCSA DB to File
Use this script to backup the full VMware VCSA 6.5 or higher datebase. The script creates a current date named folder and places a backup copy of the database in that folder.

This script will store all vCenter Username/Password and the Backup Location Password in an AES encrypted flat format to allow for future runnings of this script.

Use this script with a Windows Task Scheduler or VMware Orchestrator to schedule/automate the backup process on a timely bases.